pyproject.toml
[project]
name = "omk-python-project-template"
version = "0.1.0"
description = "A clean, structured Python project template for data and ML workflows."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]
dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.3.5",
    "hydra-core>=1.3.2",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "bump-my-version>=1.2.4",
    "cloc>=0.2.5",
    "radon>=6.0.1",
    "ruff>=0.14.5",
]
notebooks = [
    "ipykernel>=7.1.0",
    "notebook>=7.4.7",
]
test = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[tool.uv]
default-groups = [
    "dev",
    "test",
    "notebooks"
]

[project.urls]
repository = "https://github.com/OhMyKode/omk-python-project-template.git"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = ["tests/", "src/"]
markers = [
    "model: Model related tests.",
    "data: Data related tests."
]

[tool.black]
line-length = 200

[tool.ruff]
line-length = 200
include = ["*.py", "**/pyproject.toml"]
extend-include = ["*.ipynb"]

[tool.ruff.lint]
ignore = [
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "E501", # line too long, handled by black
]
select = ["E", "F", "W", "C", "B", "I001"]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv"
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pycodestyle]
max-doc-length = 200

[tool.bumpversion]
current_version = "0.1.0"
commit = false
tag = false

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "README.md"
search = 'version-{current_version}-gree'
replace = 'version-{new_version}-gree'

[[tool.bumpversion.files]]
filename = "Makefile"
search = 'version = {current_version}'
replace = 'version = {new_version}'
